[Version 1.5.0]   - Update for Bukkit 1.1-R5 and later (Including 1.2 builds)   - Removed Register, added Vault   - Added /wz reload command   - Cost is now doubles, allows non-whole numbers (5.25 for example)

+++ src/net/TheDgtl/Warpz0r/Locations.java
@@ -86,9 +86,9 @@ public class Locations {
                                              Double.parseDouble(elements[2]), 
                                              Double.parseDouble(elements[3]));
                 l.setYaw(Float.parseFloat(elements[4]));
-                int cost = -1;
+                double cost = -1;
                 if (elements.length > 6)
-                	cost = Integer.parseInt(elements[6]);
+                	cost = Double.parseDouble(elements[6]);
                 
                 List.put(elements[0].toLowerCase(), new Warp(elements[0], elements[5], l, cost));
             }
@@ -152,7 +152,7 @@ public class Locations {
         return warp.loc;
     }
     
-    public static void addWarp(Location loc, String name, int cost) {
+    public static void addWarp(Location loc, String name, double cost) {
         Locations.warps.put(name.toLowerCase(), new Warp(name, loc, cost));
         Locations.updateList();
     }
@@ -182,7 +182,7 @@ public class Locations {
         return warp.loc;
     }
     
-    public static int getWarpCost(String name) {
+    public static double getWarpCost(String name) {
     	Warp warp = Locations.warps.get(name.toLowerCase());
     	if (warp == null) return -1;
     	return warp.cost;
@@ -196,20 +196,20 @@ public class Locations {
     	public String fullName;
         public String world;
         public Location loc;
-        public int cost;
+        public double cost;
         
         Warp (String fullName, String world, Location loc) {
         	this(fullName, world, loc, -1);
         }
         
-        Warp(String fullName, String world, Location loc, int cost) {
+        Warp(String fullName, String world, Location loc, double cost) {
         	this.fullName = fullName;
             this.loc = loc;
             this.world = world;
             this.cost = cost;
         }
         
-        Warp (String fullName, Location loc, int cost) {
+        Warp (String fullName, Location loc, double cost) {
         	this.fullName = fullName;
             this.loc = loc;
             this.world = "";
